Moved azure-specific code to provider_azure.go
Called using the GroupsFetcher interface
